Fixed followingSequence()!
Just need to fix that if the user makes a mistake, it is an immediate game over and not only when the user has clicked x amount of times according to the number of items in the sequence. Also fix the animation so that it is not so confusing when the user gets the sequence right and the next sequence flashes immediately thereafter.